requirements.txt
ultralytics
torch
torchvision
pillow
numpy
opencv-python